+++ utils/include/learnopengl/animator.h
@@ -74,7 +74,7 @@ public:
 private:
 	std::vector<glm::mat4> m_FinalBoneMatrices;
 	Animation* m_CurrentAnimation;
-	float m_CurrentTime;
-	float m_DeltaTime;
+	GLdouble m_CurrentTime;
+	GLdouble m_DeltaTime;
 
 };
